expose warnings
